backend/ServicioEstadisticas/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pyodbc==5.1.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
